pyproject.toml
[project]
name = "psy-helper"
version = "0.1.0"
description = "AI-ассистент психолога: база знаний + сопровождение клиентов между сессиями"
requires-python = ">=3.11,<3.13"
dependencies = [
    "whisperx>=3.1.5",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.2.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["psy_helper*"]
